Add script to filter countries based on the input letter
The script takes a letter as an argument and filters the list of countries to select only those starting with the specified letter. It then runs a Python script 'bot_follow.py' for each filtered country, executing it a specified number of times. If there are no countries starting with the input letter, it displays a message.